move extension actions to separate file
