+++ src/components/frame/Extensions/pages/ExtensionDetailsPage/ExtensionDetailsPage.tsx
@@ -2,18 +2,16 @@ import React, { FC } from 'react';
 import { FormattedMessage, useIntl } from 'react-intl';
 import { useParams } from 'react-router-dom';
 
-import { useColonyContext, useExtensionData, useMobile } from '~hooks';
-import Button from '~shared/Extensions/Button';
+import { useColonyContext, useExtensionData } from '~hooks';
 import Icon from '~shared/Icon';
 
 import ExtensionDetails from './partials/ExtensionDetails';
-import { useExtensionDetailsPage } from './hooks';
 import Spinner from '~shared/Extensions/Spinner';
 import ThreeColumns from '~frame/Extensions/ThreeColumns';
 import Navigation from '~common/Extensions/Navigation';
 import SupportingDocuments from '~common/Extensions/SupportingDocuments';
 import ImageCarousel from '~common/Extensions/ImageCarousel';
-import { isInstalledExtensionData } from '~utils/extensions';
+import ActionButtons from '../partials/ActionButtons';
 
 const HeadingChunks = (chunks: React.ReactNode[]) => (
   <h4 className="font-semibold text-gray-900 mt-6 mb-4">{chunks}</h4>
@@ -26,10 +24,6 @@ const ExtensionDetailsPage: FC = () => {
   const { colony } = useColonyContext();
   const { extensionData } = useExtensionData(extensionId ?? '');
   const { formatMessage } = useIntl();
-  const isMobile = useMobile();
-  const { handleEnableButtonClick, handleInstallClick } = useExtensionDetailsPage();
-
-  const isExtensionInstalled = extensionData && isInstalledExtensionData(extensionData);
 
   if (!colony) {
     return null;
@@ -54,16 +48,7 @@ const ExtensionDetailsPage: FC = () => {
               <h4 className="ml-2 text-xl font-semibold text-gray-900">{formatMessage(extensionData.name)}</h4>
             </div>
             <div className="sm:ml-4">
-              {!isExtensionInstalled && (
-                <Button mode="primarySolid" isFullSize={isMobile} onClick={handleInstallClick}>
-                  <p className="text-sm font-medium">{formatMessage({ id: 'extension.installButton' })}</p>
-                </Button>
-              )}
-              {extensionData.isInitialized && !extensionData.isInitialized && (
-                <Button mode="primarySolid" isFullSize={isMobile} onClick={handleEnableButtonClick}>
-                  <p className="text-sm font-medium">{formatMessage({ id: 'extension.enableButton' })}</p>
-                </Button>
-              )}
+              <ActionButtons />
             </div>
           </div>
         }
